REFACTOR: Remove all old references from Asset Redeem Point Actor Network Service, plus modified PluginRoot's name.
Associated Issue: #5720
