+++ DAP/plugin/actor_network_service/fermat-dap-plugin-actor-network-service-redeem-point-bitdubai/src/main/java/com/bitdubai/fermat_dap_plugin/layer/actor/network/service/asset/redeem_point/developer/bitdubai/DeveloperBitDubai.java
@@ -8,7 +8,6 @@ import com.bitdubai.fermat_api.layer.all_definition.enums.CryptoCurrency;
 import com.bitdubai.fermat_api.layer.all_definition.enums.Developers;
 import com.bitdubai.fermat_api.layer.all_definition.enums.TimeFrequency;
 import com.bitdubai.fermat_api.layer.all_definition.license.PluginLicensor;
-import com.bitdubai.fermat_dap_plugin.layer.actor.network.service.asset.redeem_point.developer.bitdubai.version_1.AssetRedeemActorNetworkServicePluginRootNew;
 import com.bitdubai.fermat_dap_plugin.layer.actor.network.service.asset.redeem_point.developer.bitdubai.version_1.AssetRedeemPointActorNetworkServicePluginRoot;
 
 /**
@@ -26,7 +25,7 @@ public class DeveloperBitDubai extends AbstractPluginDeveloper implements Plugin
         try {
 
 //            this.registerVersion(new AssetRedeemPointActorNetworkServicePluginRoot());
-            this.registerVersion(new AssetRedeemActorNetworkServicePluginRootNew());
+            this.registerVersion(new AssetRedeemPointActorNetworkServicePluginRoot());
 
         } catch (CantRegisterVersionException e) {
