--- DAP/plugin/actor_network_service/fermat-dap-plugin-actor-network-service-redeem-point-bitdubai/src/main/java/com/bitdubai/fermat_dap_plugin/layer/actor/network/service/asset/redeem_point/developer/bitdubai/version_1/event_handlers/CompleteComponentConnectionRequestNotificationEventHandler.java
@@ -1,62 +1,0 @@
-package com.bitdubai.fermat_dap_plugin.layer.actor.network.service.asset.redeem_point.developer.bitdubai.version_1.event_handlers;
-
-import com.bitdubai.fermat_api.FermatException;
-import com.bitdubai.fermat_api.Service;
-import com.bitdubai.fermat_api.layer.all_definition.enums.ServiceStatus;
-import com.bitdubai.fermat_api.layer.all_definition.events.interfaces.FermatEvent;
-import com.bitdubai.fermat_api.layer.all_definition.events.interfaces.FermatEventHandler;
-import com.bitdubai.fermat_p2p_api.layer.all_definition.common.network_services.interfaces.NetworkService;
-import com.bitdubai.fermat_p2p_api.layer.all_definition.communication.events.CompleteComponentConnectionRequestNotificationEvent;
-
-/**
- * Created by franklin on 18/10/15.
- */
-public class CompleteComponentConnectionRequestNotificationEventHandler implements FermatEventHandler {
-
-    /*
-    * Represent the networkService
-    */
-    private NetworkService networkService;
-
-    /**
-     * Constructor with parameter
-     *
-     * @param networkService
-     */
-    public CompleteComponentConnectionRequestNotificationEventHandler(NetworkService networkService) {
-        this.networkService = networkService;
-    }
-
-    /**
-     * (non-Javadoc)
-     *
-     * @see FermatEventHandler#handleEvent(FermatEvent)
-     *
-     * @param platformEvent
-     * @throws Exception
-     */
-    @Override
-    public void handleEvent(FermatEvent platformEvent) throws FermatException {
-
-        System.out.println("CompleteComponentConnectionRequestNotificationEventHandler - handleEvent platformEvent ="+platformEvent );
-
-
-        if (((Service) this.networkService).getStatus() == ServiceStatus.STARTED) {
-
-            CompleteComponentConnectionRequestNotificationEvent completeComponentConnectionRequestNotificationEvent = (CompleteComponentConnectionRequestNotificationEvent) platformEvent;
-
-
-            if (completeComponentConnectionRequestNotificationEvent.getNetworkServiceTypeApplicant() ==
-                    this.networkService.getPlatformComponentProfilePluginRoot().getNetworkServiceType()){
-
-                /*
-                 *  networkService make the job
-                 */
-                this.networkService.handleCompleteComponentConnectionRequestNotificationEvent(completeComponentConnectionRequestNotificationEvent.getApplicantComponent(), completeComponentConnectionRequestNotificationEvent.getRemoteComponent());
-
-            }
-
-
-        }
-    }
-}
