--- DAP/plugin/actor_network_service/fermat-dap-plugin-actor-network-service-redeem-point-bitdubai/src/main/java/com/bitdubai/fermat_dap_plugin/layer/actor/network/service/asset/redeem_point/developer/bitdubai/version_1/event_handlers/CompleteComponentRegistrationNotificationEventHandler.java
@@ -1,53 +1,0 @@
-package com.bitdubai.fermat_dap_plugin.layer.actor.network.service.asset.redeem_point.developer.bitdubai.version_1.event_handlers;
-
-import com.bitdubai.fermat_api.FermatException;
-import com.bitdubai.fermat_api.Service;
-import com.bitdubai.fermat_api.layer.all_definition.enums.ServiceStatus;
-import com.bitdubai.fermat_api.layer.all_definition.events.interfaces.FermatEvent;
-import com.bitdubai.fermat_api.layer.all_definition.events.interfaces.FermatEventHandler;
-import com.bitdubai.fermat_p2p_api.layer.all_definition.common.network_services.interfaces.NetworkService;
-import com.bitdubai.fermat_p2p_api.layer.all_definition.communication.events.CompleteComponentRegistrationNotificationEvent;
-
-/**
- * Created by franklin on 18/10/15.
- */
-public class CompleteComponentRegistrationNotificationEventHandler implements FermatEventHandler  {
-    /*
-* Represent the networkService
-*/
-    private NetworkService networkService;
-
-    /**
-     * Constructor with parameter
-     *
-     * @param networkService
-     */
-    public CompleteComponentRegistrationNotificationEventHandler(NetworkService networkService) {
-        this.networkService = networkService;
-    }
-
-    /**
-     * (non-Javadoc)
-     *
-     * @see FermatEventHandler#handleEvent(FermatEvent)
-     *
-     * @param platformEvent
-     * @throws Exception
-     */
-    @Override
-    public void handleEvent(FermatEvent platformEvent) throws FermatException {
-
-        System.out.println("CompleteComponentRegistrationNotificationEventHandler - handleEvent platformEvent ="+platformEvent );
-
-        if (((Service) this.networkService).getStatus() == ServiceStatus.STARTED) {
-
-            CompleteComponentRegistrationNotificationEvent completeComponentRegistrationNotificationEvent = (CompleteComponentRegistrationNotificationEvent) platformEvent;
-
-            /*
-             *  networkService make the job
-             */
-            this.networkService.handleCompleteComponentRegistrationNotificationEvent(completeComponentRegistrationNotificationEvent.getPlatformComponentProfileRegistered());
-
-        }
-    }
-}
